Build fix.
svn:sync-xref-src-repo-rev: r116303

+++ src/VBox/Devices/Audio/DevHDA.cpp
@@ -1,4 +1,4 @@
-/* $Id: DevHDA.cpp 67554 2017-06-22 09:17:17Z andreas.loeffler@oracle.com $ */
+/* $Id: DevHDA.cpp 67555 2017-06-22 09:28:21Z andreas.loeffler@oracle.com $ */
 /** @file
  * DevHDA - VBox Intel HD Audio Controller.
  *
@@ -6337,7 +6337,7 @@ static int hdaLoadExecLegacy(PHDASTATE pThis, PSSMHANDLE pSSM, uint32_t uVersion
             if (RT_FAILURE(rc))
                 break;
 
-            cStreams = RT_CLAMP(cStreams, 0, HDA_MAX_STREAMS); /* Sanity. */
+            cStreams = RT_CLAMP((int32_t)cStreams, 0, HDA_MAX_STREAMS); /* Sanity. */
 
             /* Load stream states. */
             for (uint32_t i = 0; i < cStreams; i++)
@@ -6536,7 +6536,7 @@ static DECLCALLBACK(int) hdaLoadExec(PPDMDEVINS pDevIns, PSSMHANDLE pSSM, uint32
     rc = SSMR3GetU32(pSSM, &cStreams);
     AssertRC(rc);
 
-    cStreams = RT_CLAMP(cStreams, 0, HDA_MAX_STREAMS); /* Sanity. */
+    cStreams = RT_CLAMP((int32_t)cStreams, 0, HDA_MAX_STREAMS); /* Sanity. */
 
     Log2Func(("cStreams=%RU32\n", cStreams));
